SIMD optimization
build: add USE_AVX cmake argument to enable avx support
docs: add simd instruction
feat: add gemmf_32/gemm_f64
feat: test_generate_data compute the anser of M@N
test: test_MatMul dtype is now double
feat: add test_MatMulCMD to test single thread C++ matmul
feat: add identification to time elapsed

+++ Projects/Project3/tests/test_MatMul.cpp
@@ -8,12 +8,12 @@ int main() {
     using mpimath::Matrix2D;
     MPI_Init(NULL, NULL);
     MPIProcessorInfo Processor;
-    Matrix2D<int> A(9, 9, true);
+    Matrix2D<double> A(9, 9, true);
 
     LOGD("%p, A[0][0] = %d", A.pData, A[0][0]);
     A[0][0] = 1;
     A[0][1] = 2;
-    Matrix2D<int> B;
+    Matrix2D<double> B;
     B = A;
     LOGD("%p, %p", A.pData, B.pData);
